Initial version. Contains home page, view page, and edit page.
The program is contained within the Main.py file. The other files are UI files created with the pyqt5 designer. The home page and view page currently have no functions other than navigating to the other pages. However, you cannot navigation from view to edit or vice versa. Only via the home page can the program be navigated. The edit page currently has the function to add a table which included protection from any miss inputs such as duplicate tables. Various attributes with unique names, datatypes, and constraints can be added. As many commands as needed can be made in series such as creating multiple tables in a row. These are all stored in a queue (FIFO) and executed in series once the execute button is pressed. The same counts for adding multiple attributes, however, these commands are executed once a new table is created. All the attributes created before a table is created are added to that table.