Update and end of project, for now
The new update includes the ability to add entities to a specified table. The table can be chosen via a dropdown and the attributes in their dropdown are automatically updated to reflect the new table. As development progressed it became apparent that SQL has many very unique aspects. For example, there are different datatypes which have different sizes where the sizes can have different meanings (such as the size of a line of text and the size of a number). This makes it difficult and not very fun to program especially on the foundation that the code is currently on. The foundation is made assuming that the code can be automated, however, by the looks of it many pieces of code have to be written for each individual SQL function.